Move CSI and Tabix code into vcf_utils module
